+++ app/src/main/java/com/android/personalmushaf/mushafinterfaces/strategies/navigationstrategies/NavigationActivityStrategy.java
@@ -4,7 +4,7 @@ import com.android.personalmushaf.navigation.ViewPagerAdapter;
 import com.android.personalmushaf.navigation.tabs.rukucontenttab.RukuContentFragment;
 
 public interface NavigationActivityStrategy {
-    int getJuzLength(int juzNumber);
+    double getJuzLength(int juzNumber);
 
     void setViewPagerTabs(ViewPagerAdapter viewPagerAdapter, RukuContentFragment rukuContentFragment);
 }
